solvers: move to flat hierarchy
