+++ src/vqr/api.py
@@ -18,9 +18,9 @@ from vqr.vqr import (
     inversion_sampling,
 )
 from vqr.coverage import measure_coverage
-from vqr.solvers.primal.cvx import CVXVQRSolver
-from vqr.solvers.primal.pot import POTVQESolver
-from vqr.solvers.dual.regularized_lse import (
+from vqr.solvers.cvx import CVXVQRSolver
+from vqr.solvers.pot import POTVQESolver
+from vqr.solvers.regularized_lse import (
     RegularizedDualVQRSolver,
     MLPRegularizedDualVQRSolver,
 )
