+++ src/vqr/solvers/cvx.py
