+++ src/vqr/solvers/regularized_lse.py
